characterTransmitFix

+++ gamesSDLCPlusPlus CW TogDev/settings.h
@@ -53,8 +53,8 @@ struct imagePath
 
 struct point
 {
-	short p1 = 0;
-	short p2 = 0;
+	int p1 = 0;
+	int p2 = 0;
 };
